src/fiko/apps/accounts/serializers/user.py
from django.contrib.auth import get_user_model
from rest_framework import serializers

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ("id","first_name","last_name","birth_date","phone_number","invite_code","image","created_at","updated_at","date_joined","is_active")
        #fields = "__all__"

class UserUpdateSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ("first_name","last_name","birth_date","invite_code")

class UserProfilePictureSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ("image",)
